Move to util pkg
